Rename module `driver` to `stepper`
This is a follow-up to #83. I meant to include this change there, but
forgot.
